+++ packages/bridge-nest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.6
+
+### Patch Changes
+
+- neo4j link providers
+
 ## 0.3.5
 
 ### Patch Changes