+++ packages/domain-social/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.1
+
+### Patch Changes
+
+- neo4j link providers
+
+- Updated dependencies []:
+  - @volontariapp/bridge-nest@0.3.6
+
 ## 0.2.0
 
 ### Minor Changes